requirements.txt
groq
edge-tts
python-dotenv
